+++ .github/CHANGELOG.md
@@ -147,6 +147,7 @@
 1. [EFB] Added a cold temperature correction calculator - @tracernz (Mike)
 1. [A380X/MFD] Fixed PERF APPR QNH entries in inHg being displayed as hPa - @Daboss57 (Daboss57)
 1. [EFB] Add FAA and LIDO charts supplied by MSFS2024 - @tracernz (Mike)
+1. [A32NX/FWS] Add intermediate callouts to FWC - @BravoMike99 (bruno_pt99)
 
 ## 0.14.0